+++ tests/MockedData.php
@@ -24,7 +24,7 @@ class MockedData
                     'scenario' => 'manual',
                     'scope' => 'ip',
                     'type' => 'ban',
-                    'value' => Constants::IP_V4_2, ],
+                    'value' => Constants::IP_V4_2,],
             ],
             'deleted' => [],
         ],
@@ -35,7 +35,7 @@ class MockedData
                     'scenario' => 'manual',
                     'scope' => 'ip',
                     'type' => 'ban',
-                    'value' => Constants::IP_V4, ],
+                    'value' => Constants::IP_V4,],
             ],
             'deleted' => [],
         ],
@@ -46,13 +46,13 @@ class MockedData
                     'scenario' => 'manual',
                     'scope' => 'ip',
                     'type' => 'bypass',
-                    'value' => Constants::IP_V4_2, ],
+                    'value' => Constants::IP_V4_2,],
                 ['duration' => '147h',
                     'origin' => 'CAPI',
                     'scenario' => 'manual',
                     'scope' => 'ip',
                     'type' => 'ban',
-                    'value' => Constants::IP_V4_2, ],
+                    'value' => Constants::IP_V4_2,],
             ],
             'deleted' => [],
         ],
@@ -82,7 +82,7 @@ class MockedData
                     'scenario' => 'manual',
                     'scope' => 'range',
                     'type' => 'ban',
-                    'value' => Constants::IP_V4 . '/' . Constants::IP_RANGE, ],
+                    'value' => Constants::IP_V4 . '/' . Constants::IP_RANGE,],
             ],
             'deleted' => [],
         ],
@@ -93,7 +93,7 @@ class MockedData
                     'scenario' => 'manual',
                     'scope' => 'range',
                     'type' => 'ban',
-                    'value' => Constants::IP_V6 . '/' . Constants::IP_RANGE, ],
+                    'value' => Constants::IP_V6 . '/' . Constants::IP_RANGE,],
             ],
             'deleted' => [],
         ],
@@ -104,7 +104,7 @@ class MockedData
                     'scenario' => 'manual',
                     'scope' => 'range',
                     'type' => 'ban',
-                    'value' => Constants::IP_V4 . '/' . Constants::IP_RANGE, ],
+                    'value' => Constants::IP_V4 . '/' . Constants::IP_RANGE,],
             ],
             'new' => [],
         ],
@@ -219,6 +219,219 @@ class MockedData
                     'type' => 'ban',
                     'value' => 'FR',
                 ],
+            ],
+            'deleted' => [],
+        ],
+    ];
+
+    public const DECISIONS_CAPI_V3 = [
+        'new_ip_v4' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'ip',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4_2
+                        ]
+                    ]
+                ]
+            ],
+            'deleted' => [],
+        ],
+        'new_ip_v4_other' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'ip',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4
+                        ]
+                    ]
+                ]
+            ],
+            'deleted' => [],
+        ],
+        'new_ip_v4_double' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'ip',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4_2
+                        ]
+                    ]
+                ],
+                [
+                    "scenario" => "crowdsecurity/http-sensitive",
+                    "scope" => 'ip',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4_3
+                        ]
+                    ]
+                ],
+
+            ],
+            'deleted' => [],
+        ],
+        'deleted_ip_v4' => [
+            'deleted' => [
+                [
+                    "scope" => 'ip',
+                    "decisions" =>
+                    [
+                        Constants::IP_V4_2,
+                    ]
+                ]
+            ],
+            'new' => [],
+        ],
+        'new_ip_v4_range' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'range',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4 . '/' . Constants::IP_RANGE
+                        ]
+                    ]
+                ]
+            ],
+            'deleted' => [],
+        ],
+        'new_ip_v6_range' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'range',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V6 . '/' . Constants::IP_RANGE
+                        ]
+                    ]
+                ]
+            ],
+            'deleted' => [],
+        ],
+        'delete_ip_v4_range' => [
+            'deleted' => [
+                [
+                    "scope" => 'range',
+                    "decisions" =>
+                        [
+                            Constants::IP_V4 . '/' . Constants::IP_RANGE,
+                        ]
+                ]
+            ],
+            'new' => [],
+        ],
+        'ip_v4_multiple' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'ip',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4_2
+                        ]
+                    ]
+                ],
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'range',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4 . '/' . Constants::IP_RANGE
+                        ]
+                    ]
+                ],
+                [
+                    "scenario" => "crowdsecurity/http-sensitive",
+                    "scope" => 'ip',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4
+                        ]
+                    ]
+                ],
+            ],
+            'deleted' => [],
+        ],
+        'ip_v4_multiple_bis' => [
+            'deleted' => [
+                [
+                    "scope" => 'ip',
+                    "decisions" =>
+                        [
+                            Constants::IP_V4,
+                        ]
+                ]
+            ],
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'ip',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4_3
+                        ]
+                    ]
+                ]
+            ],
+        ],
+        'ip_v4_remove_unknown' => [
+            'deleted' => [
+                [
+                    "scope" => 'do-not-know-delete',
+                    "decisions" =>
+                        [
+                            Constants::IP_V4,
+                        ]
+                ]
+            ],
+            'new' => [],
+        ],
+        'ip_v4_store_unknown' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'do-not-know-store',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => Constants::IP_V4
+                        ]
+                    ]
+                ]
+            ],
+            'deleted' => [],
+        ],
+        'country_ban' => [
+            'new' => [
+                [
+                    "scenario" => "crowdsecurity/http-backdoors-attempts",
+                    "scope" => 'Country',
+                    "decisions" => [
+                        [
+                            'duration' => '147h',
+                            'value' => 'FR'
+                        ]
+                    ]
+                ]
             ],
             'deleted' => [],
         ],