+++ tests/scripts/refresh-decisions-capi.php
@@ -8,14 +8,14 @@ use CrowdSec\RemediationEngine\CacheStorage\Memcached;
 use CrowdSec\RemediationEngine\CacheStorage\PhpFiles;
 use CrowdSec\RemediationEngine\CacheStorage\Redis;
 use CrowdSec\RemediationEngine\CapiRemediation;
-use CrowdSec\Common\Logger\FileLog;
+use CrowdSec\Common\Logger\ConsoleLog;
 
 // Init  logger
-$logger = new FileLog(['debug_mode' => true], 'remediation-engine-logger');
+$logger = new ConsoleLog();
 // Init client
 $clientConfigs = [
     'machine_id_prefix' => 'remediationtest',
-    'scenarios' => ['crowdsecurity/http-sensitive-files'],
+    'scenarios' => ["crowdsecurity/http-backdoors-attempts","crowdsecurity/http-bad-user-agent"],
 ];
 $capiClient = new Watcher($clientConfigs, new FileStorage(__DIR__), null, $logger);
 // Init PhpFiles cache storage